api/pyproject.toml
[project]
name = "humidity-tracker-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "pyserial==3.5",
    "psycopg[binary]>=3.1",
]
